still working on the same fix

+++ distdrone/imgproc/__init__.py
@@ -46,12 +46,13 @@ def findfaceswithtrigger(profile='picluster',threshold=15,n=100,serial='no',show
 			file=open('mypythonpid','r')
 			mypid=cPickle.load(file)
 			file.close()
+		except:
+			mypid=False
+		if mypid:
 			try:
 				os.kill(mypid,0)
 			except:
 				mypid=False
-		except:
-			mypid=False
 		return mypid
 	def getrunstate():
 		import cPickle
